Switching Pagar.me Checkout activity
